refactor logging utils
