+++ src/logging/config.rs
@@ -327,42 +327,12 @@ impl LogConfig {
 
         // Validate file size format
         if self.outputs.file.enabled {
-            self.parse_file_size(&self.outputs.file.max_size)?;
+            super::utils::parse_file_size(&self.outputs.file.max_size)?;
         }
 
         Ok(())
     }
 
-    /// Parse file size string (e.g., "10MB", "1GB") to bytes
-    fn parse_file_size(&self, size_str: &str) -> Result<u64, ConfigError> {
-        let size_str = size_str.to_uppercase();
-
-        if let Some(num_str) = size_str.strip_suffix("GB") {
-            let num: u64 = num_str
-                .parse()
-                .map_err(|_| ConfigError::InvalidFileSize(size_str.clone()))?;
-            Ok(num * 1024 * 1024 * 1024)
-        } else if let Some(num_str) = size_str.strip_suffix("MB") {
-            let num: u64 = num_str
-                .parse()
-                .map_err(|_| ConfigError::InvalidFileSize(size_str.clone()))?;
-            Ok(num * 1024 * 1024)
-        } else if let Some(num_str) = size_str.strip_suffix("KB") {
-            let num: u64 = num_str
-                .parse()
-                .map_err(|_| ConfigError::InvalidFileSize(size_str.clone()))?;
-            Ok(num * 1024)
-        } else if let Some(num_str) = size_str.strip_suffix("B") {
-            num_str
-                .parse()
-                .map_err(|_| ConfigError::InvalidFileSize(size_str.clone()))
-        } else {
-            // Default to bytes if no suffix
-            size_str
-                .parse()
-                .map_err(|_| ConfigError::InvalidFileSize(size_str.clone()))
-        }
-    }
 }
 
 /// Configuration errors
